+++ src/dublib/telebot_utils/master/decorators.py
@@ -3,7 +3,7 @@ from time import sleep
 
 from telebot import apihelper
 
-def ignore_frecuency_errors(function):
+def ignore_frequency_errors(function):
 	"""
 	Декоратор. Игнорирует ошибки частоты запросов, автоматически выжидая необходимый интервал.